refactor(evaluation): Move EvaluationResult interface to types file and enhance TestCase interface
- Moved the EvaluationResult interface to a new file `evaluation.types.ts` for better modularity and organization.
- Enhanced the TestCase interface to extend from CoreTestCase, adding new properties for improved clarity and structure.
- Removed the obsolete `evaluationTypes.ts` file to streamline the codebase.

This change improves the maintainability of the evaluation module by consolidating type definitions and ensuring a clearer structure for test case evaluations.
