+++ src/evaluation/types/evaluation.types.ts
@@ -1,18 +1,24 @@
 import { BaseMessage } from '../../core/types/message.types';
+import { TestCase as CoreTestCase } from '../../core/types/message.types';
 
 /**
  * Represents a test case for agent evaluation.
  */
-export interface TestCase {
+export interface TestCase extends CoreTestCase {
+  /**
+   * Unique identifier for the test case.
+   */
+  id: string;
+
+  /**
+   * Name of the test case.
+   */
+  name: string;
+
   /**
    * The sequence of messages that form the test case.
    */
-  messages: BaseMessage[];
-
-  /**
-   * The expected output message from the agent.
-   */
-  expectedOutput: BaseMessage;
+  messageBlocks: BaseMessage[];
 }
 
 /**
@@ -38,4 +44,45 @@ export interface TestCaseResult {
    * Time taken to execute the test case in milliseconds.
    */
   executionTime: number;
+}
+
+/**
+ * Represents the complete result of evaluating a test case, including the test case itself
+ * and the agent's response.
+ */
+export interface EvaluationResult {
+  /**
+   * Unique identifier of the test case.
+   */
+  testCaseId: string;
+
+  /**
+   * Whether the test case execution was successful.
+   */
+  success: boolean;
+
+  /**
+   * Error message if the test case failed.
+   */
+  error?: string;
+
+  /**
+   * Score between 0 and 1 indicating how well the agent performed.
+   */
+  score: number;
+
+  /**
+   * Time taken to execute the test case in milliseconds.
+   */
+  executionTime: number;
+
+  /**
+   * The response from the agent.
+   */
+  response?: string;
+
+  /**
+   * The original test case that was evaluated.
+   */
+  testCase: TestCase;
 }
